ManagedObject#saveInBackground performs a block on another thread.
MR doesn't set `NSManagedObjectContext.contextForCurrentThread` to the
`localContext` that's yielded to the block, which means you **have** to
specify the context for each insert.

We make this a bit easier by storing the `localContext` as a Thread var
for the duration of the block, which is used by `ManagedObject.new`.